requirements.txt
absl-py==1.4.0
alabaster==0.7.13
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
base58==2.1.1
Bashutils==0.0.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
cmake==3.26.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
docutils==0.18.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.3.3
fonttools==4.39.4
future==0.18.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpc-interceptor==0.14.2
grpcio==1.48.2
grpcio-tools==1.48.2
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
ifaddr==0.2.0
imageio==2.26.0
imageio-ffmpeg==0.4.8
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
incremental==22.10.0
ipaddress==1.0.23
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==3.0
joblib==1.2.0
jsonschema==4.17.3
jupyter_core==5.3.0
keras==2.11.0
kiwisolver==1.4.4
lamda==3.138
layers==0.1.5
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
moviepy==1.0.3
multiaddr==0.0.9
nbformat==5.7.3
netaddr==0.8.0
netifaces==0.11.0
ntplib==0.4.0
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python~=4.5.2.52
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pafy==0.5.5
pandas==1.5.3
parso==0.8.3
peerpy==1.2.2
pickleshare==0.7.5
Pillow==9.4.0
pkcs1==0.9.6
platformdirs==3.1.1
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.18.0
pydot==1.4.2
pydotplus==2.0.2
pydub==0.25.1
Pygments==2.14.0
pylab-sdk==1.3.2
PyNaCl==1.5.0
pyp2p==0.8.2
pyparsing==3.0.9
pyreadline==2.1
Pyro4==4.82
pyroute2==0.7.6
pyrsistent==0.19.3
python-dateutil==2.8.2
PythonInterfaces==0.0.2
pytube==15.0.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sequential==1.0.0
serpent==1.41
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.6
Sphinx==6.2.1
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow~=2.10.1
tensorflow-docs @ git+https://github.com/tensorflow/docs@63191f2e07e92aefc9da30f442465e55e7eb7d07
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
typing_extensions==4.5.0
urllib3==1.26.15
varint==1.0.2
wcwidth==0.2.6
Werkzeug==1.0.1
win-inet-pton==1.1.0
wrapt==1.15.0
youtube-dl==2021.12.17
zeroconf==0.47.4
zipp==3.15.0
zope.interface==6.0
cryptography~=40.0.2
marshmallow~=3.19.0
schedule~=1.2.0
paho-mqtt~=1.6.1
remotezip~=0.12.1
seaborn~=0.12.2